Cleanup UnpackEvent
Cleanup of the UnpackEvent class: Remove dependency from the
Processor class FRSUnpackProc and from the EventSource.
The event is built via the API of Go4 framework that call
BuildEvent in the Processors class. There are clear difference
between the Event class that are the data structure that will be
forwarded to the other processor class and/or to the ROOT tree.
Therefore the Fill member function is remove, the Init is
simplified.