Cleanup UnpackProc
Add BuildEvent member function, which is part of the API of Go4
framework. The BuildEvent correspond to the previously named
member function FRSUnpack. Now instead to call FRSUnpack within
the UnpackEvent, now the framework call BuildEvent by itself.
Add FillHistogram member function in which all histogram filling
is conditionally done. Create Event_Extract member function to
is the actual code that unpack the data from the lmd file into the
UnpackEvent. Event_Extract is called inside the BuildEvent function
depending of the trigger flag. In such a way, the BuildEvent only
take care of the event header and of the trigger flags and delegate
the actual data extraction to a separated function for more clarity.
TFRSParameter is now part of the UnpackProc class in order to avoid
multiple assignment during the event loop, now it is created and
assigned in the constructor.