Cleanup Calibr step similar to Unpack and Sort
Cleanup of Calibr step: In Event remove dependency to the Proc.
In Proc change to BuildEvent Go4 API instead of FRSCalibr.
Additionally, the InitProcessor is call at construction time.
It insures that the object of the CalibrProc class is initialized
at its creation. Also the parameters (FRS, MW, TPC, SI) were
retrieved at each call of each dedicated analysis, now they are
attributes of the CalibrProc class and are only retrieved once in
the constructor. The names Create_*_Data were change to
Create_*_Hist since those member function are to create histograms.