Add flexible setup initialization in FRSAnalysis
The file setup.C was used for the parameters initialization.
It was call inside TFRSAnalysis and send to be interpreted by
ROOT processLine. Add some flexibility were one setup file per
experiment now be written and be under the version control of git
repository. An configuration file is now necessary in the directory
of the libGo4UserAnalysis.so where go4 starts. The file "config.txt"
just hold the name of experiment that is concatenated to the
setup file name. For example : setup_s411.C, the name experiment is
then s411 and inside of the config.txt there has to be a line :
"NameExperiment s411"
New setup file can then created with any name that follow this
manner : setup_*.C, where the * is the name. Please each new setup
file has to be added the git repository for control any changes.