Implement MRtof and FRS data synchronization
Implement the synchronization of the MRtof data with the lastest
FRS analysis data. When the flag of the FRS data is there, save
the data event within the FRS process step. When the flag of the
MRtof is there, the MRTOF information is transmitted to the
User event structure then the lastest saved FRS analysis data is
transmitted to the same User event structure.
The SetValid flag of the User event object will tell to the go4
framework to flush the current event to tree or not.
In this way the final root TTree will only included synchronized
event of MRtof data with FRS data